+++ src/venmo/index.js
@@ -20,6 +20,7 @@ var VERSION = process.env.npm_package_version;
  * @param {Client} [options.client] A {@link Client} instance.
  * @param {string} [options.authorization] A tokenizationKey or clientToken. Can be used in place of `options.client`.
  * @param {boolean} [options.allowNewBrowserTab=true] This should be set to false if your payment flow requires returning to the same tab, e.g. single page applications. Doing so causes {@link Venmo#isBrowserSupported|isBrowserSupported} to return true only for mobile web browsers that support returning from the Venmo app to the same tab.
+ * @param {boolean} [options.allowWebviews=true] This should be set to false if your payment flow does not occur from within a webview that you control. Doing so causes {@link Venmo#isBrowserSupported|isBrowserSupported} to return true only for mobile web browsers that are not webviews.
  * @param {boolean} [options.ignoreHistoryChanges=false] When the Venmo app returns to the website, it will modify the hash of the url to include data about the tokenization. By default, the SDK will put the state of the hash back to where it was before the change was made. Pass `true` to handle the hash change instead of the SDK.
  * @param {string} [options.profileId] The Venmo profile ID to be used during payment authorization. Customers will see the business name and logo associated with this Venmo profile, and it will show up in the Venmo app as a "Connected Merchant". Venmo profile IDs can be found in the Braintree Control Panel. Omitting this value will use the default Venmo profile.
  * @param {string} [options.deepLinkReturnUrl] An override for the URL that the Venmo iOS app opens to return from an app switch.
@@ -90,6 +91,7 @@ function create(options) {
  * @function isBrowserSupported
  * @param {object} [options] browser support options:
  * @param {boolean} [options.allowNewBrowserTab=true] This should be set to false if your payment flow requires returning to the same tab, e.g. single page applications.
+ * @param {boolean} [options.allowWebviews=true] This should be set to false if your payment flow does not occur from within a webview that you control.
  * @example
  * if (braintree.venmo.isBrowserSupported()) {
  *   // set up Venmo
@@ -97,6 +99,12 @@ function create(options) {
  * @example <caption>Explicitly require browser support returning to the same tab</caption>
  * if (braintree.venmo.isBrowserSupported({
  *   allowNewBrowserTab: false
+ * })) {
+ *   // set up Venmo
+ * }
+ * @example <caption>Explicitly set webviews as disallowed browsers</caption>
+ * if (braintree.venmo.isBrowserSupported({
+ *   allowWebviews: false
  * })) {
  *   // set up Venmo
  * }