+++ test/venmo/unit/is-browser-supported.js
@@ -8,9 +8,11 @@ const { isBrowserSupported } = require('../../../src/venmo/shared/supports-venmo
 describe('isBrowserSupported', () => {
   beforeEach(() => {
     jest.spyOn(browserDetection, 'isIosSafari').mockReturnValue(false);
+    jest.spyOn(browserDetection, 'isIosWebview').mockReturnValue(false);
     jest.spyOn(browserDetection, 'isIos').mockReturnValue(false);
     jest.spyOn(browserDetection, 'isChrome').mockReturnValue(false);
     jest.spyOn(browserDetection, 'isAndroid').mockReturnValue(false);
+    jest.spyOn(browserDetection, 'isAndroidWebview').mockReturnValue(false);
     jest.spyOn(browserDetection, 'isSamsungBrowser').mockReturnValue(false);
     jest.spyOn(browserDetection, 'isMobileFirefox').mockReturnValue(false);
   });
@@ -19,6 +21,30 @@ describe('isBrowserSupported', () => {
     browserDetection.isIosSafari.mockReturnValue(true);
 
     expect(isBrowserSupported()).toBe(true);
+  });
+
+  it('returns true for iOS Safari when webviews are dissallowed', () => {
+    browserDetection.isIosSafari.mockReturnValue(true);
+
+    expect(isBrowserSupported({
+      allowWebviews: false
+    })).toBe(true);
+  });
+
+  it('returns true for iOS webview', () => {
+    browserDetection.isIosSafari.mockReturnValue(true);
+    browserDetection.isIosWebview.mockReturnValue(true);
+
+    expect(isBrowserSupported()).toBe(true);
+  });
+
+  it('returns false for iOS webview when configured to dissallow webviews', () => {
+    browserDetection.isIosSafari.mockReturnValue(true);
+    browserDetection.isIosWebview.mockReturnValue(true);
+
+    expect(isBrowserSupported({
+      allowWebviews: false
+    })).toBe(false);
   });
 
   it('returns true for iOS Chrome', () => {
@@ -33,6 +59,24 @@ describe('isBrowserSupported', () => {
     browserDetection.isChrome.mockReturnValue(true);
 
     expect(isBrowserSupported()).toBe(true);
+  });
+
+  it('returns true for Android webview', () => {
+    browserDetection.isAndroid.mockReturnValue(true);
+    browserDetection.isChrome.mockReturnValue(true);
+    browserDetection.isAndroidWebview.mockReturnValue(true);
+
+    expect(isBrowserSupported()).toBe(true);
+  });
+
+  it('returns false for Android webview when configured', () => {
+    browserDetection.isAndroid.mockReturnValue(true);
+    browserDetection.isChrome.mockReturnValue(true);
+    browserDetection.isAndroidWebview.mockReturnValue(true);
+
+    expect(isBrowserSupported({
+      allowWebviews: false
+    })).toBe(false);
   });
 
   it('returns true for Samsung Browser', () => {